Updated MFCC F32 Neon version. Removed some compilation warnings in the tests

+++ Testing/Include/Tests/UnaryTestsF32.h
@@ -69,6 +69,6 @@ class UnaryTestsF32:public Client::Suite
             float64_t *outa;   
             float64_t *outb;
 
-            float32_t snrRel,snrAbs;
+            double snrRel,snrAbs;
 
     };
